Refactor RichTextEditor, fix slider styling, and adjust theme tones
- Refactored Icons.tsx to use map-based lookup. - Extracted RichTextEditor sub-components to components/rich-text/. - Fixed slider styling in index.html. - Adjusted Warm and Cool light theme palettes.
